Word LM decoding (and various other changes)

+++ egs/discophone/v1/local/chain_multilang/tuning/run_tdnn_1g.sh
@@ -36,7 +36,7 @@ gp_recog="${gp_langs}"
 # (some of which are also used in this script directly).
 stage=0
 stop_stage=500
-num_epochs=4
+num_epochs=8
 get_egs_stage=0
 nj=30
 nj_align_fmllr_lats=100
@@ -127,7 +127,8 @@ local/chain_multilang/run_ivector_common.sh --stage $stage --stop-stage $stop_st
   --gp-recog $gp_recog \
   --gmm $gmm \
   --num-threads-ubm $num_threads_ubm \
-  --nnet3-affix "$nnet3_affix"
+  --nnet3-affix "$nnet3_affix" \
+  --data-aug-suffix "$data_aug_suffix"
 
 data_dir=$train_set
 lang_name=universal
@@ -244,10 +245,6 @@ EOF
 fi
 
 if [ $stage -le 18 ] && [ $stop_stage -gt 18 ]; then
-  #if [[ $(hostname -f) == *.clsp.jhu.edu ]] && [ ! -d $dir/egs/storage ]; then
-  #  utils/create_split_dir.pl \
-  #   /export/b0{5,6,7,8}/$USER/kaldi-data/egs/aishell-$(date +'%m_%d_%H_%M')/s5c/$dir/egs/storage $dir/egs/storage
-  #fi
   echo "exit stage is $train_exit_stage"
   steps/nnet3/chain/train.py --stage $train_stage --exit-stage $train_exit_stage \
     --cmd "$decode_cmd" \
@@ -278,7 +275,7 @@ if [ $stage -le 18 ] && [ $stop_stage -gt 18 ]; then
     --dir $dir || exit 1
 fi
 
-if [ $stage -le 19 ]; then
+if [ $stage -le 19 ] && [ $stop_stage -gt 19 ]; then
   echo "$0: creating high-resolution MFCC features for the test data"
   mfccdir=mfcc${data_aug_suffix}_hires
 
@@ -301,11 +298,12 @@ if [ $stage -le 19 ]; then
   wait
 fi
 
-if [ $stage -le 20 ]; then
-  echo "$0: creating high-resolution MFCC features for the test data"
+if [ $stage -le 20 ] && [ $stop_stage -gt 20 ]; then
+  echo "$0: extracting i-vectors for the test data"
   for data_dir in ${recog_set}; do
     nspk=$(wc -l <data/${data_dir}_hires/spk2utt)
     steps/online/nnet2/extract_ivectors_online.sh --cmd "$train_cmd" --nj "${nspk}" \
+      --use-vad true \
       data/${data_dir}_hires exp/nnet3/universal/extractor \
       exp/nnet3/universal/ivectors_${data_dir}_hires
   done
@@ -313,7 +311,7 @@ fi
 
 LMTYPE=phn_bg_multi
 
-if [ $stage -le 21 ]; then
+if [ $stage -le 21 ] && [ $stop_stage -gt 21 ]; then
   # The reason we are using data/lang here, instead of $lang, is just to
   # emphasize that it's not actually important to give mkgraph.sh the
   # lang directory with the matched topology (since it gets the
@@ -328,7 +326,7 @@ if [ $stage -le 21 ]; then
     $tree_dir $tree_dir/graph_${LMTYPE} || exit 1
 fi
 
-if [ $stage -le 22 ]; then
+if [ $stage -le 22 ] && [ $stop_stage -gt 22 ]; then
   frames_per_chunk=$(echo $frames_per_eg | cut -d, -f1)
   rm $dir/.error 2>/dev/null || true
 
@@ -351,13 +349,70 @@ if [ $stage -le 22 ]; then
       for lmtype in $LMTYPE; do
         steps/nnet3/decode.sh \
           --use-gpu $gpu_decode \
-          --acwt 1.0 --post-decode-acwt 10.0 \
           --extra-left-context 0 --extra-right-context 0 \
           --extra-left-context-initial 0 \
           --extra-right-context-final 0 \
-          --frames-per-chunk $frames_per_chunk \
-          --nj $nj_decode --cmd "$cmd_decode" --num-threads $nt_decode \
+          --num-threads $nt_decode \
           --online-ivector-dir exp/nnet3/universal/ivectors_${data}_hires \
+          --acwt 1.0 --post-decode-acwt 10.0 \
+          --frames-per-chunk $frames_per_chunk \
+          --nj $nj_decode --cmd "$cmd_decode" \
+          $tree_dir/graph_${lmtype} data/${data}_hires ${dir}/decode_${lmtype}_${data_affix} || exit 1
+      done
+    ) || touch $dir/.error &
+  done
+  wait
+  [ -f $dir/.error ] && echo "$0: there was a problem while decoding" && exit 1
+fi
+
+langdir=data/lang_universal_test
+
+if [ $stage -le 23 ] && [ $stop_stage -gt 23 ]; then
+  # The reason we are using data/lang here, instead of $lang, is just to
+  # emphasize that it's not actually important to give mkgraph.sh the
+  # lang directory with the matched topology (since it gets the
+  # topology file from the model).  So you could give it a different
+  # lang directory, one that contained a wordlist and LM of your choice,
+  # as long as phones.txt was compatible.
+
+  utils/lang/check_phones_compatible.sh \
+    data/lang_chain/lang_${lang_name}/phones.txt $langdir/phones.txt
+  utils/mkgraph.sh \
+    --self-loop-scale 1.0 $langdir \
+    $tree_dir $tree_dir/graph_words || exit 1
+fi
+
+if [ $stage -le 24 ] && [ $stop_stage -gt 24 ]; then
+  frames_per_chunk=$(echo $frames_per_eg | cut -d, -f1)
+  rm $dir/.error 2>/dev/null || true
+
+  for data in $recog_set; do
+    (
+      data_affix=$(echo $data | sed s/eval_//)
+      nspk=$(wc -l <data/${data}_hires/spk2utt)
+
+      if $gpu_decode; then
+        nj_decode=1
+        nt_decode=16
+        cmd_decode="$cuda_cmd"
+      else
+        nj_decode=$nspk
+        nt_decode=1
+        cmd_decode="$decode_cmd"
+      fi
+
+      # (pzelasko): Eventually we'll have more LM types here, for now it's just one
+      for lmtype in words; do
+        steps/nnet3/decode.sh \
+          --use-gpu $gpu_decode \
+          --extra-left-context 0 --extra-right-context 0 \
+          --extra-left-context-initial 0 \
+          --extra-right-context-final 0 \
+          --num-threads $nt_decode \
+          --online-ivector-dir exp/nnet3/universal/ivectors_${data}_hires \
+          --acwt 1.0 --post-decode-acwt 10.0 \
+          --frames-per-chunk $frames_per_chunk \
+          --nj $nj_decode --cmd "$cmd_decode" \
           $tree_dir/graph_${lmtype} data/${data}_hires ${dir}/decode_${lmtype}_${data_affix} || exit 1
       done
     ) || touch $dir/.error &
